[Rubocop] updated to rubocop-0.49.1 (#25)
* New cops Rails/ApplicationJob and Rails/ApplicationRecord
It requires some support in code base, but always a good idea

* AlignParameters, MultilineMethodCallIndentation, MultilineOperationIndentation, SpaceInLambdaLiteral and SpaceInsideBrackets cops moved From Style/ to Layout/

* Enabled Layout/MultilineMethodCallIndentation
Most violations of this cop i found in our ode were awfull

* New cops Lint/RescueType and Lint/ScriptPermission
They are good and we had no violations

* New cop Performance/Caller
It's ruby 2.* today

* Finnaly disabled RSpec/FilePath
Rubocop fixed a bug when it could not be disabled

* New cop Style/FormatStringToken
Experimental. `%{}` is suggested by very much of guides, but `%<>s` seems better. At least, that guides never told about `%<>s`

* New cop Style/MultipleComparison
It suggest Array#include?, but we can use `Object#in?` in rails projects. Only one violation found!

* Disabled new cop Style/YodaCondition
This is not cop you are looking for